Add commandline to csv files
When saving a csv file add commandline args to file.

`# warp put --access-key=*REDACTED* --secret-key=*REDACTED* --md5=true --duration=1s`

The reason we don't use the raw args is that this will not be correct on remote hosts.

We filter access/secret keys as can be seen in the example.

Unlisted args have default value.