[get] bench object from listing the bucket
when doing a get bench, during the prepare phase, instead of putting
objects in the bucket use the objects already in there. This can be used
to bench a specific use case with an already filled in bucket. It can
also be used to speed up get tests by always using the same objects and
gain the prepare phase.

New options:
--get-from-listing-bucket: to activate the listing bucket (current
  behaviour remains)
--listing-bucket-max-objects: to limit the number of object to use
--listing-bucket-prefix: to use a specific prefix during listing
--listing-bucket-no-recursive: disable recursive listing