[get] bench object from listing the bucket (#250)
when doing a get bench, during the prepare phase, instead of putting
objects in the bucket use the objects already in there. This can be used
to bench a specific use case with an already filled in bucket. It can
also be used to speed up get tests by always using the same objects and
gain the prepare phase.

Using `--list-existing` will list at most `--objects` from the bucket and download them instead
of uploading random objects (set it to 0 to use all object from the listing).
Listing is restricted to `--prefix` if it is set and recursive listing can be disabled by setting `--list-flat`