Fix hang at end of distributed warp with influxdb (#340)
Fixed issue where distributed warp hangs at end

When running distributed warp with influxdb, the app would always
hang at the end of each run. This is because the channel used
to send operations to Influx isn't used on the initiator process
and thus, never gets closed. The app then hangs on the global
wait group for the influx goroutine to complete.