Add a format for object-size proviging size buckets
Allows a coma-separated set of colon-separated pairs, describing buckets
and their respective weights. This format triggers an option that
performs a weighted random number generation when a new object is
created.

E.g.: `4096:10740,8192:1685,16384:1623` will trigger objects whose size
will be chosen between 0 and 4096 with a weight of 10740, between 4096
and 8192 with a weight of 1685, or between 8192 and 16384 with a weight
of 1623.