Add new "randomLite" generator
The existing default generator ('random') uses AES-CBC to generate truly random object data. Under very heavy load, this could become a bottleneck and was sometimes crashing. For our purposes when dedupe is not enabled, we can use the same random data just so it's not compressible. A fixed set of random data is placed in a buffer (2MiB+1 by default) and accessed through the circular buffer interface. We use an odd number so objects get slightly different offsets and thus, different ETAG values.