Highlight the danger of data being cleaned (#426)
* Highlight the danger of data being cleaned

I made a big mistake locally, that I suspect others might fall victim to. As such, I propose to highlight the danger of data being wiped with the warning-format. If that is to bold, then I suggest the note variant. Either way, I would argue something needs to be done to warn users of potential data loss, and not in the form of a last sentence in an arbitrary paragraph with an italicized "not".

* Update README.md

---------

Co-authored-by: Klaus Post <klauspost@gmail.com>