Add --prefixes flag for GCS and pre-sharded bucket
There are 2 use-cases for a static set of object prefixes ("folders"):
 1. GCS makes deleting "folders" difficult, at least with the
    heirarchical optimization feature enabled, and S3 API delete calls
    can't delete them.  So if Warp generates a fresh one for every
    thread that ever runs, it makes cleaning up the bucket super
    tedious, even though it can have zero objects.
 2. Testing a major CSP's object storage involved having them pre-shard
    the bucket on known prefixes to ensure a valid test.  We needed Warp
    to support this, so patched in the support.  (That patch was
    hackier, and this is a better version of it, though it worked fine
    for our usage.)

This change preserves the behavior of the --prefix and --noprefix flags.

We add a new flag, "--prefixes value", which is mutually exclusive with
--prefix and whose value is a comma-separated list of one or more static
prefix values.  The case of --prefix and --prefixes with one value
specified are identical (and in fact, --prefix is now implemented as a
single-element --prefixes internally).

When multiple prefixes are given with --prefixes, each object within
each thread has a static prefix chosen from the list in a round-robin
fashion.  Also, the listings for --list-existing include all the
prefixes, "chaining" the prefix listings together.

Use-case 1 above will be satisfied by specifying --prefixes
with the desired count of top-level "folders" and --noprefix.

Use-case 2 can be satisfied with just --prefixes (though --noprefix
doesn't hurt it).

The interactions of the flags are complicated, so a new unit test was
added to cover the various combinations.

bench.{delete,get,stat} now use a helper function with an options struct
to reduce logic and code duplication.

Cleaned up separation of concerns between generator.{generator,random},
ensuring all prefix generation logic is in generator.random.