Fix get to allow --objects 0 (per README)
    Using `--list-existing` will list at most `--objects` from the
    bucket and download them instead of uploading random objects
    (set it to 0 to use all object from the listing).

The intent of --objects with --list-existing for the "get" command was
for the special value 0 to represent "list all and use all".  This is
correctly implemented for "delete" but "get" was disallowing this
combination.

This fix allows --objects 0 with --list-existing while ensuring the
object count is >= 1 without --list-existing and still not negative when
--list-existing is supplied.