Distributed delete runs no longer overlap (#441)
Previously, when a distributed (more than one client) delete benchmark
deleted objects with --list-existing, all clients would try to delete
all objects.  This resulted in a lot of spurious errors.

This patch makes each client only delete a partition (based on a hash of
the object path) of the full object list.

This patch does NOT try to address the server and all clients trying to
run Cleanup() for non-delete benchmark types.  That is to say, this
patch does not mess with Common.deleteAllInBucket().

Co-authored-by: Klaus Post <klauspost@gmail.com>